modules: disaggregate display.py
